% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workinghard.R
\name{workinghard}
\alias{workinghard}
\title{Title}
\usage{
workinghard(n, d, t)
}
\arguments{
\item{t}{Hard work indicator boolean (TRUE or FALSE)}

\item{N}{Name - must be character class}

\item{a}{day - must be numeric class}
}
\value{
Returns a statement of acknowledgement
}
\description{
Title
}
\examples{
workinghard("Alex",1,TRUE)
}
